refactor(aiget/server): 提取 AgentStreamProcessor 从 AgentService
- 将流式事件处理逻辑（consumeStreamEvents, convertRunEventToSSE）提取到独立的 AgentStreamProcessor 服务
- AgentService 从 1029 行减少到 873 行
- AgentStreamProcessor 212 行，职责单一：处理 Agent 运行时流式事件并转换为 SSE 格式
- 更新测试用例以适配新的依赖注入
